# Shift limits from row to col in rowvec/colvec banded_broadcast