--- a/Project.toml
+++ b/Project.toml
@@ -6,12 +6,14 @@ version = "0.17.10"
 ArrayLayouts = "4c555306-a7a7-4459-81d9-ec55ddd5c99a"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 
 [compat]
 Aqua = "0.5"
 ArrayLayouts = "0.8.14"
 FillArrays = "0.13"
+MacroTools = "0.5"
 julia = "1.6"
 
 [extras]