--- a/src/generic/broadcast.jl
+++ b/src/generic/broadcast.jl
@@ -855,7 +855,7 @@ function checkzerobands(dest, f, (A,B)::Tuple{AbstractMatrix,AbstractMatrix})
     B_l, B_u = bandwidths(B)
     l, u = max(A_l,B_l), max(A_u,B_u)
 
-    @switchlimits for j = 1:n
+    for j = 1:n
         for k = max(1,j-u) : min(j-d_u-1,m)
             iszero(f(A[k,j], B[k,j])) || throw(BandError(dest,b))
         end