The various banded broadcast methods have loops of the form
```julia
for j=1:n
        for k = max(1,j-d_u):min(j-u-1,m)
            inbands_setindex!(dest, z, k, j)
        end
end
```
This involves `max` and `min` calls for each `j`, which adds an `O(n)` overhead. It seems this isn't negligible, and shifting the limits from `k` to `j` improves performance by a significant extent. Doing this manually is tedious, error-prone, and hard to maintain, so I've written a macro to do this.

With arrays
```julia
julia> A = BandedMatrix(0=>Float64[1:3000;]); D = Diagonal(A);

julia> A2 = brand(3000, 3000, (200,200)); D2 = Diagonal(A2);
```
Runtimes:
| Operation | master | PR |
| --- | :-: |  :-: |
|`A * D` | 54.493 μs | 13.244 μs |
|`D * A` | 49.778 μs | 45.225 μs |
|`A2 * D2` | 53.917 μs | 12.976 μs |
|`D2 * A2` | 49.282 μs | 44.779 μs |

This only really improves things if `A` also has few bands, so we may add a special path for that. If the number of bands is high, this will likely make things worse due to cache misses. I see the times being roughly equal for tridiagonal matrices. Perhaps we should only use this for diagonal matrices, in which case this is a clear improvement.

However, in the diagonal case, a dot product is much more performant, so this is perhaps unnecessary.

I've added [MacroTools](https://github.com/FluxML/MacroTools.jl) as a dependency to make life simpler. This adds 15.0 ms to the package load time, which is negligible compared to the other dependencies. This should also be maintained, as it's a part of the FluxML organization.