@dlfivefifty 2023-02-07
I’m wary of adding macros which are very hard to maintain….